Move ctxerr.Visited in its own package
