+++ internal/visited/visited.go
@@ -4,7 +4,7 @@
 // This source code is licensed under the BSD-style license found in the
 // LICENSE file in the root directory of this source tree.
 
-package ctxerr
+package visited
 
 import (
 	"reflect"
